renamed RCT proto to avoid namespace clash
