+++ infrastructure/src/main/java/org/corfudb/infrastructure/remotecorfutable/RemoteCorfuTableRequestHandler.java
@@ -18,13 +18,13 @@ import static org.corfudb.protocols.service.CorfuProtocolMessage.getHeaderMsg;
 import static org.corfudb.protocols.service.CorfuProtocolMessage.getResponseMsg;
 import org.corfudb.runtime.proto.service.CorfuMessage.RequestMsg;
 import org.corfudb.runtime.proto.service.CorfuMessage.ResponseMsg;
-import org.corfudb.runtime.proto.service.RemoteCorfuTable.RemoteCorfuTableContainsKeyRequestMsg;
-import org.corfudb.runtime.proto.service.RemoteCorfuTable.RemoteCorfuTableContainsValueRequestMsg;
-import org.corfudb.runtime.proto.service.RemoteCorfuTable.RemoteCorfuTableGetRequestMsg;
-import org.corfudb.runtime.proto.service.RemoteCorfuTable.RemoteCorfuTableMultiGetRequestMsg;
-import org.corfudb.runtime.proto.service.RemoteCorfuTable.RemoteCorfuTableRequestMsg;
-import org.corfudb.runtime.proto.service.RemoteCorfuTable.RemoteCorfuTableScanRequestMsg;
-import org.corfudb.runtime.proto.service.RemoteCorfuTable.RemoteCorfuTableSizeRequestMsg;
+import org.corfudb.runtime.proto.service.RemoteCorfuTableMessages.RemoteCorfuTableContainsKeyRequestMsg;
+import org.corfudb.runtime.proto.service.RemoteCorfuTableMessages.RemoteCorfuTableContainsValueRequestMsg;
+import org.corfudb.runtime.proto.service.RemoteCorfuTableMessages.RemoteCorfuTableGetRequestMsg;
+import org.corfudb.runtime.proto.service.RemoteCorfuTableMessages.RemoteCorfuTableMultiGetRequestMsg;
+import org.corfudb.runtime.proto.service.RemoteCorfuTableMessages.RemoteCorfuTableRequestMsg;
+import org.corfudb.runtime.proto.service.RemoteCorfuTableMessages.RemoteCorfuTableScanRequestMsg;
+import org.corfudb.runtime.proto.service.RemoteCorfuTableMessages.RemoteCorfuTableSizeRequestMsg;
 
 import javax.annotation.Nonnull;
 import java.util.List;
